+++ src/main/java/com/realstate/services/TenantService.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.realstate.domains.Tenant;
+import com.realstate.exceptions.TenantDoesNotExistException;
 import com.realstate.repositories.TenantRepository;
 
 @Service
@@ -15,28 +16,50 @@ public class TenantService {
 	
 	@Autowired
 	private TenantRepository tenantRepository;
-
+	
+	public Tenant getNew(String fullName, String dni, String phone, String description) {
+		Tenant tenant = new Tenant(null, fullName, dni, phone, description);
+		return tenantRepository.insert(tenant);
+	}
+	
 	public List<Tenant> findAll() {
 		return tenantRepository.findAll();
 	}
 	
-	public Optional<Tenant> findById(String tenantId) {
-		return tenantRepository.findById(new ObjectId(tenantId));
+	public Tenant findById(String tenantId) throws TenantDoesNotExistException {
+		Optional<Tenant> optionalTenant = tenantRepository.findById(new ObjectId(tenantId));
+		if (optionalTenant.isPresent()) {
+			return optionalTenant.get();
+		} else {
+			throw new TenantDoesNotExistException();
+		}
 	}
 	
 	public List<Tenant> findByDni(String tenantDni) {
 		return tenantRepository.findAllByDni(tenantDni);
 	}
 	
+	public boolean existById(String tenantId) {
+		return tenantRepository.existsById(new ObjectId(tenantId));
+	}
+	
 	public Tenant insert(Tenant newTenant) {
 		return tenantRepository.insert(newTenant);
 	}
 	
-	public Tenant update(Tenant tenant) {
-		return tenantRepository.save(tenant);
+	public Tenant update(Tenant tenant) throws TenantDoesNotExistException {
+		if (tenantRepository.existsById(new ObjectId(tenant.getTenantId()))) {
+			return tenantRepository.save(tenant);
+		} else {
+			throw new TenantDoesNotExistException();
+		}
 	}
 	
-	public void delete(Tenant tenant) {
-		tenantRepository.delete(tenant);
+	public void delete(Tenant tenant) throws TenantDoesNotExistException {
+		if (tenantRepository.existsById(new ObjectId(tenant.getTenantId()))) {
+			tenantRepository.delete(tenant);
+		} else {
+			throw new TenantDoesNotExistException();
+		}
 	}
 }
